+++ org/mozilla/jss/nss/Buffer.c
@@ -5,7 +5,7 @@
 
 #include "jssutil.h"
 #include "BufferProxy.h"
-#include "buffer.h"
+#include "j_buffer.h"
 
 #include "_jni/org_mozilla_jss_nss_Buffer.h"
