+++ org/mozilla/jss/ssl/javax/j_buffer.h
